removed some outdated files, moved files
