+++ lib/widgets/receive_form.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:seeds/providers/services/navigation_service.dart';
-import 'package:seeds/screens/app/wallet/amount_field.dart';
+import 'package:seeds/widgets/amount_field.dart';
 import 'package:seeds/widgets/main_button.dart';
 import 'package:seeds/i18n/wallet.i18n.dart';
